chore(*): Test refactoing - fixing xit-ed test. Refactoring kbnav to use promises.
